+++ videos/tests/test_views.py
@@ -27,26 +27,41 @@ class TestVideoList(Python2CountEqualMixin, RequestTestCase):
 
 
 class TestVideoDetail(RequestTestCase):
-    def test_get_detail(self):
-        video = factories.VideoFactory.create()
-        source = factories.SourceFactory.create(video=video)
-        chapter = ChapterFactory.create(video=video)
-        speaker = SpeakerFactory.create()
-        video.speakers.add(speaker)
+    def setUp(self):
+        self.video = factories.VideoFactory.create()
+        self.source = factories.SourceFactory.create(video=self.video)
+        self.chapter = ChapterFactory.create(video=self.video)
+        self.speaker = SpeakerFactory.create()
+        self.video.speakers.add(self.speaker)
 
+    def test_get_detail(self):
         view = views.VideoDetail.as_view()
-        response = view(self.create_request(auth=False), slug=video.slug)
+        response = view(self.create_request(auth=False), slug=self.video.slug)
 
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.context_data['object'], video)
+        self.assertEqual(response.context_data['object'], self.video)
 
         rendered_content = response.render().content.decode()
 
-        self.assertIn(video.title, rendered_content)
-        self.assertIn(source.get_absolute_url(), rendered_content)
-        self.assertIn(source.get_type_display(), rendered_content)
-        self.assertIn('href="#{}"'.format(chapter.seconds), rendered_content)
-        self.assertIn(speaker.name, rendered_content)
+        self.assertIn(self.video.title, rendered_content)
+        self.assertIn(self.source.get_absolute_url(), rendered_content)
+        self.assertIn(self.source.get_type_display(), rendered_content)
+        self.assertIn('href="#{}"'.format(self.chapter.seconds), rendered_content)
+        self.assertIn(self.speaker.name, rendered_content)
+
+    def test_get_content_data(self):
+        view = views.VideoDetail()
+        view.object = self.video
+        context_data = view.get_context_data()
+        expected = {
+            'video': self.video,
+            'view': view,
+            'object': self.video,
+            'sources': [self.source],
+            'chapters': [self.chapter],
+            'speakers': [self.speaker],
+        }
+        self.assertCountEqual(context_data, expected)
 
 
 class TestVideoListLatest(RequestTestCase):