+++ videos/views.py
@@ -16,3 +16,14 @@ class VideoListLatest(VideoList):
 class VideoDetail(DetailView):
     template_name = 'videos/detail.html'
     model = Video
+
+    def get_context_data(self, **kwargs):
+        extensions = ('chapters', 'speakers')
+        for extension in extensions:
+            manager = getattr(self.object, extension, None)
+            if manager:
+                kwargs.update({extension: manager.all()})
+        return super(VideoDetail, self).get_context_data(
+            sources=self.object.sources.all(),
+            **kwargs
+        )